Move alchemy recipe-related classes to new subpackage
