+++ src/main/java/com/mraof/minestuck/Minestuck.java
@@ -2,7 +2,7 @@ package com.mraof.minestuck;
 
 import com.mraof.minestuck.advancements.MSCriteriaTriggers;
 import com.mraof.minestuck.alchemy.GristTypes;
-import com.mraof.minestuck.alchemy.generator.recipe.InterpreterSerializers;
+import com.mraof.minestuck.alchemy.recipe.generator.recipe.InterpreterSerializers;
 import com.mraof.minestuck.block.MSBlocks;
 import com.mraof.minestuck.blockentity.MSBlockEntityTypes;
 import com.mraof.minestuck.util.MSParticleType;
