+++ src/main/java/com/mraof/minestuck/alchemy/recipe/CombinerWrapper.java
@@ -1,4 +1,4 @@
-package com.mraof.minestuck.alchemy;
+package com.mraof.minestuck.alchemy.recipe;
 
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
