+++ src/main/java/com/mraof/minestuck/alchemy/recipe/ItemCombiner.java
@@ -1,5 +1,6 @@
-package com.mraof.minestuck.alchemy;
+package com.mraof.minestuck.alchemy.recipe;
 
+import com.mraof.minestuck.alchemy.recipe.CombinationMode;
 import net.minecraft.world.Container;
 
 public interface ItemCombiner extends Container
