+++ src/main/java/com/mraof/minestuck/alchemy/recipe/generator/GeneratedCostProvider.java
@@ -1,4 +1,4 @@
-package com.mraof.minestuck.alchemy.generator;
+package com.mraof.minestuck.alchemy.recipe.generator;
 
 import net.minecraft.world.item.Item;
