+++ src/main/java/com/mraof/minestuck/alchemy/recipe/generator/GenerationContext.java
@@ -1,4 +1,4 @@
-package com.mraof.minestuck.alchemy.generator;
+package com.mraof.minestuck.alchemy.recipe.generator;
 
 import com.mraof.minestuck.alchemy.GristSet;
 import net.minecraft.world.item.Item;
