+++ src/main/java/com/mraof/minestuck/alchemy/recipe/generator/recipe/InterpreterSerializers.java
@@ -1,4 +1,4 @@
-package com.mraof.minestuck.alchemy.generator.recipe;
+package com.mraof.minestuck.alchemy.recipe.generator.recipe;
 
 import com.mraof.minestuck.Minestuck;
 import net.minecraft.resources.ResourceLocation;
