+++ src/main/java/com/mraof/minestuck/blockentity/machine/TotemLatheBlockEntity.java
@@ -6,9 +6,9 @@ import com.mraof.minestuck.block.machine.TotemLatheBlock;
 import com.mraof.minestuck.blockentity.MSBlockEntityTypes;
 import com.mraof.minestuck.item.MSItems;
 import com.mraof.minestuck.alchemy.AlchemyHelper;
-import com.mraof.minestuck.alchemy.CombinationMode;
-import com.mraof.minestuck.alchemy.CombinationRecipe;
-import com.mraof.minestuck.alchemy.CombinerWrapper;
+import com.mraof.minestuck.alchemy.recipe.CombinationMode;
+import com.mraof.minestuck.alchemy.recipe.CombinationRecipe;
+import com.mraof.minestuck.alchemy.recipe.CombinerWrapper;
 import com.mraof.minestuck.blockentity.ItemStackBlockEntity;
 import com.mraof.minestuck.util.ColorHandler;
 import com.mraof.minestuck.util.WorldEventUtil;
