+++ src/main/java/com/mraof/minestuck/network/EditmodeDragPacket.java
@@ -1,7 +1,7 @@
 package com.mraof.minestuck.network;
 
 import com.mraof.minestuck.MinestuckConfig;
-import com.mraof.minestuck.alchemy.GristCost;
+import com.mraof.minestuck.alchemy.recipe.GristCost;
 import com.mraof.minestuck.alchemy.GristSet;
 import com.mraof.minestuck.alchemy.GristTypes;
 import com.mraof.minestuck.computer.editmode.*;
